feat(playback): add 1-second empty time before playback
Introduces a one-second empty period before the chart begins.

This change prevents potential video artifacts from appearing at the start of playback or upon restarting the chart. The video surface is now cleared with an opaque black color and will only render when the playback time is positive.

The initial playback time is now set to -1000ms to accommodate this new pre-roll period.